pyproject.toml
[tool.poetry]
name = "gsp-rl"
version = "1.3.0"
description = "Stable with config support"
authors = ["jdbloom"]
readme = "README.md"
packages = [{include = "gsp_rl"}]

[tool.poetry.dependencies]
python = "^3.10"
torch = "^2.1.2"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
pandas = "^2.1.4"
SciPy = "^1.11.4"
black = "^23.12.1"
mypy = "^1.8.0"
pylint = "^3.0.3"
zmq = "^0.0.0"
gymnasium = {extras = ["box2d"], version = "^0.29.1"}
tensorboard = "^2.16.2"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=gsp_rl --cov-report=term-missing --cov-fail-under=65 -m 'not slow'"
markers = [
    "slow: convergence tests (deselect with '-m \"not slow\"')",
]
